Fix StyleCop warnings in TimexUtility
