+++ .NET/Microsoft.Recognizers.Text.DateTime/Utilities/TimexUtility.cs
@@ -13,15 +13,15 @@ namespace Microsoft.Recognizers.Text.DateTime
 
         private static readonly HashSet<char> NumberComponents = new HashSet<char>()
         {
-            '0','1','2','3','4','5','6','7','8','9','.'
+            '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '.',
         };
 
         private static readonly Dictionary<DatePeriodTimexType, string> DatePeriodTimexTypeToTimexSuffix = new Dictionary<DatePeriodTimexType, string>()
         {
-            {DatePeriodTimexType.ByDay, Constants.TimexDay },
-            {DatePeriodTimexType.ByWeek, Constants.TimexWeek },
-            {DatePeriodTimexType.ByMonth, Constants.TimexMonth },
-            {DatePeriodTimexType.ByYear, Constants.TimexYear }
+            { DatePeriodTimexType.ByDay, Constants.TimexDay },
+            { DatePeriodTimexType.ByWeek, Constants.TimexWeek },
+            { DatePeriodTimexType.ByMonth, Constants.TimexMonth },
+            { DatePeriodTimexType.ByYear, Constants.TimexYear },
         };
 
         public static string GenerateCompoundDurationTimex(Dictionary<string, string> unitToTimexComponents, IImmutableDictionary<string, long> unitValueMap)
@@ -35,7 +35,7 @@ namespace Microsoft.Recognizers.Text.DateTime
             {
                 var timexComponent = unitToTimexComponents[unitList[i]];
 
-                // The Time Duration component occurs first time, 
+                // The Time Duration component occurs first time,
                 if (!isTimeDurationAlreadyExist && IsTimeDurationTimex(timexComponent))
                 {
                     timexBuilder.Append($"{Constants.TimeTimexPrefix}{GetDurationTimexWithoutPrefix(timexComponent)}");
@@ -50,20 +50,9 @@ namespace Microsoft.Recognizers.Text.DateTime
             return timexBuilder.ToString();
         }
 
-        private static bool IsTimeDurationTimex(string timex)
-        {
-            return timex.StartsWith($"{Constants.GeneralPeriodPrefix}{Constants.TimeTimexPrefix}");
-        }
-
-        private static string GetDurationTimexWithoutPrefix(string timex)
-        {
-            // Remove "PT" prefix for TimeDuration, Remove "P" prefix for DateDuration
-            return timex.Substring(IsTimeDurationTimex(timex) ? 2 : 1);
-        }
-
         public static string GenerateDatePeriodTimex(DateObject begin, DateObject end, DatePeriodTimexType timexType, DateObject alternativeBegin = default(DateObject), DateObject alternativeEnd = default(DateObject))
         {
-            var equalDurationLength = ((end - begin) == (alternativeEnd - alternativeBegin));
+            var equalDurationLength = (end - begin) == (alternativeEnd - alternativeBegin);
 
             if (alternativeBegin.IsDefaultValue() || alternativeEnd.IsDefaultValue())
             {
@@ -86,7 +75,7 @@ namespace Microsoft.Recognizers.Text.DateTime
                         unitCount = (((end.Year - begin.Year) * 12) + (end.Month - begin.Month)).ToString();
                         break;
                     default:
-                        unitCount = ((end.Year - begin.Year) + (end.Month - begin.Month) / 12.0).ToString(CultureInfo.InvariantCulture);
+                        unitCount = ((end.Year - begin.Year) + ((end.Month - begin.Month) / 12.0)).ToString(CultureInfo.InvariantCulture);
                         break;
                 }
             }
@@ -152,8 +141,8 @@ namespace Microsoft.Recognizers.Text.DateTime
                 }
             }
 
-            return Constants.GeneralPeriodPrefix + 
-                   (isLessThanDay ? Constants.TimeTimexPrefix : string.Empty) + 
+            return Constants.GeneralPeriodPrefix +
+                   (isLessThanDay ? Constants.TimeTimexPrefix : string.Empty) +
                    number.ToString(CultureInfo.InvariantCulture) + unitStr;
         }
 
@@ -249,10 +238,10 @@ namespace Microsoft.Recognizers.Text.DateTime
                     result.EndHour = 23;
                     result.EndMin = 59;
                     break;
-               default:
+                default:
                     break;
             }
-         
+
             return result;
         }
 
@@ -289,10 +278,10 @@ namespace Microsoft.Recognizers.Text.DateTime
 
         public static RangeTimexComponents GetRangeTimexComponents(string rangeTimex)
         {
-            rangeTimex = rangeTimex.Replace("(", "").Replace(")", "");
+            rangeTimex = rangeTimex.Replace("(", string.Empty).Replace(")", string.Empty);
             var components = rangeTimex.Split(new string[] { "," }, StringSplitOptions.RemoveEmptyEntries);
             var result = new RangeTimexComponents();
-            
+
             if (components.Length == 3)
             {
                 result.BeginTimex = components[0];
@@ -313,16 +302,16 @@ namespace Microsoft.Recognizers.Text.DateTime
         {
             return timex.Replace(Constants.TimexFuzzyYear, context.Year.ToString("D4"));
         }
-    }
 
-    public class RangeTimexComponents
-    {
-        public string BeginTimex;
+        private static bool IsTimeDurationTimex(string timex)
+        {
+            return timex.StartsWith($"{Constants.GeneralPeriodPrefix}{Constants.TimeTimexPrefix}");
+        }
 
-        public string EndTimex;
-
-        public string DurationTimex;
-
-        public bool IsValid = false;
+        private static string GetDurationTimexWithoutPrefix(string timex)
+        {
+            // Remove "PT" prefix for TimeDuration, Remove "P" prefix for DateDuration
+            return timex.Substring(IsTimeDurationTimex(timex) ? 2 : 1);
+        }
     }
 }
